+++ docs/changelog.rst
@@ -25,6 +25,9 @@ would have also been included in the 1.2 line.
 Changelog
 =========
 * :feature:`540` Add support for dotted python paths to the fabric -f cli option
+* :feature:`474` `~fabric.tasks.execute` now allows you to access the executed
+  task's return values, by itself returning a dictionary whose keys are the
+  host strings executed against.
 * :bug:`487` Overhauled the regular expression escaping performed in
   `~fabric.contrib.files.append` and `~fabric.contrib.files.contains` to try
   and handle more corner cases. Thanks to Neilen Marais for the patch.